Refactor notification_store canister and update time handling
- Removed unused dependencies `serde_json` and `futures` from `Cargo.toml` and `Cargo.lock`.
- Updated the `created_at` field in `NotificationData` to use `SystemTime` for better timestamp representation.
- Replaced `ic_cdk::api::time()` with `get_current_system_time_from_ic()` in relevant functions for consistent time handling.
- Added a `post_upgrade` function to prune notifications older than 30 days using the new time handling method.
