+++ src/canister/notification_store/Cargo.toml
@@ -17,9 +17,6 @@ ic-cdk-macros = { workspace = true }
 ic-stable-structures = { workspace = true }
 ciborium = { workspace = true }
 shared_utils = { workspace = true }
-serde_json = { workspace = true }
-serde_json_any_key = "2.0.0"
-futures = { workspace = true }
 ic-cdk-timers = { workspace = true }
 
 [dev-dependencies]
