+++ src/canister/notification_store/src/lib.rs
@@ -1,10 +1,11 @@
 use candid::Principal;
 use ic_cdk::caller;
 use ic_cdk_macros::export_candid;
-use shared_utils::canister_specific::individual_user_template::types::error::NotificationStoreError;
+use shared_utils::canister_specific::notification_store::types::error::NotificationStoreError;
 use shared_utils::canister_specific::notification_store::types::notification::{Notification, NotificationData, NotificationType};
 use ic_stable_structures::memory_manager::{MemoryId, MemoryManager, VirtualMemory};
 use ic_stable_structures::{DefaultMemoryImpl, StableBTreeMap};
+use shared_utils::common::utils::system_time::get_current_system_time_from_ic;
 use std::cell::RefCell;
 use std::time::Duration;
 
@@ -30,7 +31,7 @@ fn add_notification(user: Principal, notification_type: NotificationType) -> Res
     }
     CANISTER_DATA.with(|map| {
         let next_id = NEXT_ID.with(|id| *id.borrow_mut());
-        map.borrow_mut().insert(user, Notification (vec![NotificationData { notification_id: next_id, payload: notification_type, read: false, created_at: ic_cdk::api::time() }]));
+        map.borrow_mut().insert(user, Notification (vec![NotificationData { notification_id: next_id, payload: notification_type, read: false, created_at: get_current_system_time_from_ic()}]));
         NEXT_ID.with(|id| *id.borrow_mut() += 1);
     });
 
@@ -66,9 +67,9 @@ fn get_notifications(user: Principal, limit: usize, offset: usize) -> Vec<Notifi
 fn init() {
 
     // pruning notifications older than 30 days
-    ic_cdk_timers::set_timer(Duration::from_secs(60 * 60 * 24 * 30), move || {
+    ic_cdk_timers::set_timer_interval(Duration::from_secs(60 * 60 * 24 * 30), move || {
         CANISTER_DATA.with_borrow_mut(|map|{
-            let now = ic_cdk::api::time();
+            let now = get_current_system_time_from_ic();
 
             // Collecting the user principals first to avoid borrowing issues while mutating the map
             let users: Vec<Principal> = map
@@ -89,4 +90,29 @@ fn init() {
     });
 }
 
+#[ic_cdk_macros::post_upgrade]
+pub fn post_upgrade(){
+        // pruning notifications older than 30 days
+        ic_cdk_timers::set_timer_interval(Duration::from_secs(60 * 60 * 24 * 30), move || {
+            CANISTER_DATA.with_borrow_mut(|map|{
+                let now = get_current_system_time_from_ic();
+    
+                // Collecting the user principals first to avoid borrowing issues while mutating the map
+                let users: Vec<Principal> = map
+                    .iter()
+                    .map(|(user, _)| user)
+                    .collect();
+    
+                for user in users {
+                    if let Some(mut notifications) = map.get(&user) {
+                        notifications
+                            .0
+                            .retain(|n| n.created_at > now);
+    
+                        map.insert(user, notifications);
+                    }
+                }
+            })
+        });
+}
 export_candid!();
