+++ src/lib/shared_utils/src/canister_specific/notification_store/types/mod.rs
@@ -1,1 +1,2 @@
 pub mod notification;
+pub mod error;
